move to components and add tabs
